Refactor the project into an MVC pattern
