+++ src/GDF/AllPath.java
@@ -2,6 +2,9 @@ package GDF;
 
 import java.util.ArrayList;
 
+import GDF.models.Graph;
+import GDF.models.Vertex;
+
 public class AllPath {
     public AllPath(Graph graph) {
         super();
