+++ src/GDF/MyMain.java
@@ -10,7 +10,7 @@ public class MyMain extends Application {
 
 	@Override
 	public void start(Stage stage) throws Exception {
-		Parent root = FXMLLoader.load(getClass().getResource("Main.fxml"));
+		Parent root = FXMLLoader.load(getClass().getResource("/GDF/views/Main.fxml"));
 		stage.setTitle("Graph Path Finder");
 		stage.setScene(new Scene(root, 800, 500));
 		stage.show();
