+++ src/GDF/views/About.fxml
@@ -9,7 +9,7 @@
 <?import javafx.scene.text.Font?>
 <?import javafx.scene.web.WebView?>
 
-<Pane xmlns="http://javafx.com/javafx/16" xmlns:fx="http://javafx.com/fxml/1" fx:controller="GDF.InformationSceneController">
+<Pane xmlns="http://javafx.com/javafx/16" xmlns:fx="http://javafx.com/fxml/1" fx:controller="GDF.controllers.InformationSceneController">
    <children>
       <TabPane stylesheets="@design.css" tabClosingPolicy="UNAVAILABLE">
          <tabs>
