+++ src/GDF/views/Draw.fxml
@@ -26,7 +26,7 @@
 <?import javafx.scene.paint.Color?>
 <?import javafx.scene.text.Font?>
 
-<VBox maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="800.0" prefWidth="1200.0" styleClass="parent" stylesheets="@design.css" xmlns="http://javafx.com/javafx/16" xmlns:fx="http://javafx.com/fxml/1" fx:controller="GDF.DrawSceneController">
+<VBox maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="800.0" prefWidth="1200.0" styleClass="parent" stylesheets="@design.css" xmlns="http://javafx.com/javafx/16" xmlns:fx="http://javafx.com/fxml/1" fx:controller="GDF.controllers.DrawSceneController">
    <children>
       <HBox alignment="CENTER_LEFT" prefHeight="100.0" prefWidth="200.0" styleClass="main-pane" stylesheets="@design.css" VBox.vgrow="ALWAYS">
          <effect>
@@ -58,7 +58,7 @@
                                  <graphic>
                                     <ImageView fitHeight="20.0" fitWidth="20.0" pickOnBounds="true" preserveRatio="true">
                                        <image>
-                                          <Image url="@../icon/AddVertices.png" />
+                                          <Image url="@../../icon/AddVertices.png" />
                                        </image>
                                     </ImageView>
                                  </graphic></ToggleButton>
@@ -66,7 +66,7 @@
                                  <graphic>
                                     <ImageView fitHeight="20.0" fitWidth="20.0" pickOnBounds="true" preserveRatio="true">
                                        <image>
-                                          <Image url="@../icon/AddEdge.png" />
+                                          <Image url="@../../icon/AddEdge.png" />
                                        </image>
                                     </ImageView>
                                  </graphic>
@@ -75,7 +75,7 @@
                                  <graphic>
                                     <ImageView fitHeight="20.0" fitWidth="20.0" pickOnBounds="true" preserveRatio="true">
                                        <image>
-                                          <Image url="@../icon/Move.png" />
+                                          <Image url="@../../icon/Move.png" />
                                        </image>
                                     </ImageView>
                                  </graphic>
@@ -84,7 +84,7 @@
                                  <children>
                                     <ImageView fitHeight="20.0" fitWidth="20.0" pickOnBounds="true" preserveRatio="true">
                                        <image>
-                                          <Image url="@../icon/Delete.png" />
+                                          <Image url="@../../icon/Delete.png" />
                                        </image>
                                     </ImageView>
                                  </children>
@@ -96,7 +96,7 @@
                                  <children>
                                     <ImageView fitHeight="20.0" fitWidth="20.0" pickOnBounds="true" preserveRatio="true">
                                        <image>
-                                          <Image url="@../icon/Option.png" />
+                                          <Image url="@../../icon/Option.png" />
                                        </image>
                                     </ImageView>
                                  </children>
@@ -172,7 +172,7 @@
                                  <graphic>
                                     <ImageView fitHeight="30.0" fitWidth="30.0" pickOnBounds="true" preserveRatio="true">
                                        <image>
-                                          <Image url="@../icon/Run.png" />
+                                          <Image url="@../../icon/Run.png" />
                                        </image>
                                     </ImageView>
                                  </graphic>
@@ -184,7 +184,7 @@
                                  <children>
                                     <ImageView fitHeight="30.0" fitWidth="30.0" pickOnBounds="true" preserveRatio="true">
                                        <image>
-                                          <Image url="@../icon/Stop.png" />
+                                          <Image url="@../../icon/Stop.png" />
                                        </image>
                                     </ImageView>
                                  </children>
@@ -246,7 +246,7 @@
                            <children>
                               <ImageView fitHeight="40.0" fitWidth="40.0" pickOnBounds="true" preserveRatio="true">
                                  <image>
-                                    <Image url="@../icon/Logo.png" />
+                                    <Image url="@../../icon/Logo.png" />
                                  </image>
                               </ImageView>
                            </children>
